BasePresenterProtocol implemented
