+++ getOfferOfDream/Sources/TabBarViews/Favorite/View/FavoriteItemViewInput.swift
