--- getOfferOfDream/Sources/TabBarViews/FavoriteItem/Presenter/FavoriteItemViewDelegate.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-protocol FavoriteItemViewDelegate: AnyObject {
-    func viewDidLoad()
-}
