+++ getOfferOfDream/Sources/TabBarViews/Feedback/View/FeedbackItemViewController.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+// MARK: - FeedbackItemViewController
+class FeedbackItemViewController: UIViewController {
+
+    // MARK: - Overrides Methods
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .link
+    }
+}
