--- getOfferOfDream/Sources/TabBarViews/FeedbackItem/View/FeedbackItemViewController.swift
@@ -1,12 +1,0 @@
-import UIKit
-
-// MARK: - FeedbackItemViewController
-class FeedbackItemViewController: UIViewController {
-
-    // MARK: - Overrides Methods
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .link
-    }
-
-}
