+++ getOfferOfDream/Sources/TabBarViews/Main/QuestionDetail/Presenter/QuestionDetailPresenterDelegate.swift
