+++ getOfferOfDream/Sources/TabBarViews/Main/QuestionDetail/View/QuestionDetailViewController.swift
@@ -1,0 +1,9 @@
+import UIKit
+import GetOfferUI
+
+// MARK: -
+final class QuestionDetailViewController: UIViewController {
+
+    // MARK: - Public Properties
+    let tableView = UITableView(frame: .zero, style: .plain)
+}
