--- getOfferOfDream/Sources/TabBarViews/MainItem/QuestionDetail/View/QuestionDetailViewController.swift
@@ -1,4 +1,0 @@
-import UIKit
-import GetOfferUI
-
-// MARK: -
