--- getOfferOfDream/Sources/TabBarViews/MainItem/Questions/Presenter/QuestionsPresenterProtocol.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-protocol QuestionsPresenterProtocol {
-    func viewDidLoad()
-}
