--- getOfferOfDream/Sources/TabBarViews/MainItem/Themes/Presenter/MainItemPresenterProtocol.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-protocol MainItemPresenterProtocol: AnyObject {
-    func viewDidLoad()
-}
