+++ getOfferOfDream/Sources/TabBarViews/Quiz/View/QuizItemViewController.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+// MARK: - QuizItemViewController
+class QuizItemViewController: UIViewController {
+
+    // MARK: - Overrides Methods
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .systemTeal
+    }
+}
