--- getOfferOfDream/Sources/TabBarViews/QuizItem/View/QuizItemViewController.swift
@@ -1,12 +1,0 @@
-import UIKit
-
-// MARK: - QuizItemViewController
-class QuizItemViewController: UIViewController {
-
-    // MARK: - Overrides Methods
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .systemTeal
-    }
-
-}
