+++ getOfferOfDream/Sources/TabBarViews/Shared/BasePresenterProtocol.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol BasePresenterProtocol: AnyObject {
+    func viewDidLoad()
+}
