+++ getOfferOfDream/Sources/TabBarViews/Shared/BaseTableManagerDelegate.swift
@@ -1,0 +1,6 @@
+import UIKit
+
+protocol BaseTableManagerDelegate: AnyObject {
+    func update()
+    func setup(tableView: UITableView)
+}
