--- getOfferOfDream/Sources/TabBarViews/Shared/TableManagerDelegate.swift
@@ -1,6 +1,0 @@
-import UIKit
-
-protocol TableManagerDelegate: AnyObject {
-    func update()
-    func setup(tableView: UITableView)
-}
